fix(lists): changed behaviour of backspace and lists
Previously when we had a list like the following, Where
[] is the current selection:
```

* []a
```
and we pressed backspace, this would be the result:
```
a
```

This commit changes the behavior of pressing backspace
at the beginning of lists, so that each backspace
will remove one list level until the list element
is removed for that piece of text.

This was the preceeding linebreak is never removed,
and it will hopefully help users who are not very
adept at using using tab. Also works better on mobile.