fix(lists): change behavior of delete into lists
This commit fixes the case of selection being
on a blank line preceeding a list item, and pressing
delete into the list item.

There were two separate issues in this case.

1. When having the following, where [] is the current selection:
```

[]
* a
```
and pressing delete, then the linebreaks would be merged
into the beginning of the list item, like so:

```
*

  a
```

2. When a block had linebreaks at the end, and those linebreaks
was adjacent to a list block which we deleted into:
```
foobar

[]
* a
```

Pressing delete in this case would remove both the linebreak
and the list item, like so:
```
foobar

a
```